requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
backports.zoneinfo==0.2.1
boto3==1.26.65
botocore==1.29.65
certifi==2022.12.7
charset-normalizer==3.0.1
configparser==5.3.0
Django==4.1.5
django-ckeditor==6.5.1
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==2.0.0
django-minio-storage==0.3.10
django-storages==1.13.2
fontawesomefree==6.4.0
frozenlist==1.3.3
idna==3.4
isodate==0.6.1
jmespath==1.0.1
lxml==4.9.2
minio==6.0.2
multidict==6.0.4
Pillow==9.4.0
platformdirs==3.0.0
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==0.10.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
yarl==1.8.2
zeep==4.2.1
